+++ src/device/analog/simulator.h
@@ -14,8 +14,8 @@
 
 cq_status simulate_pulse(channel *ch, pulse *pulse, analog_qreg *qreg, cq_hamiltonian *hamiltonian);
 int simulate_capture(channel *ch, analog_qreg *qreg);
-void sync_simulator(cq_hamiltonian *hamiltonian, int qreg_id);
-void print_statevec(int qreg_id);
-void init_simulator_qreg(int qreg_id, int num_qubits);
-void reset_simulator_qreg(int qreg_id);
+void sync_simulator(cq_hamiltonian *hamiltonian, ptrdiff_t qreg_id);
+void print_statevec(ptrdiff_t qreg_id);
+void init_simulator_qreg(ptrdiff_t qreg_id, ptrdiff_t num_qubits);
+void reset_simulator_qreg(ptrdiff_t qreg_id);
 #endif // CQ_ANALOG_SIMULATOR_H
